requirements.txt
# Deep Learning Core
torch>=1.9.0
torchvision>=0.10.0

# Data Processing
numpy>=1.19.0
pillow>=8.0.0

# Configuration
pyyaml>=5.4.0

# Training
tqdm>=4.60.0

# Visualization
matplotlib>=3.3.0

# Data Augmentation (optional but recommended)
albumentations>=1.0.0

# Data Preparation (for toolkits)
kagglehub
nibabel
